Public Version of Foxes And Hounds
Simulation of foxes and hounds on a field. Hounds eat foxes (and will
replace a fox with a new hound if two hounds are adjacent to the fox
being eaten), and can starve to death. Foxes can procreate, if left
next to each other for one “timestep” without a hound nearby. This
simulation includes very basic graphics, or can be run command line.